+++ .i18nrc.json
@@ -19,6 +19,7 @@
     "kbnVislibVisTypes": "src/legacy/core_plugins/kbn_vislib_vis_types",
     "kibana_react": "src/legacy/core_plugins/kibana_react",
     "kibana-react": "src/plugins/kibana_react",
+    "kibana_utils": "src/plugins/kibana_utils",
     "navigation": "src/legacy/core_plugins/navigation",
     "newsfeed": "src/plugins/newsfeed",
     "regionMap": "src/legacy/core_plugins/region_map",
