+++ src/legacy/core_plugins/kibana/public/visualize/editor/editor.js
@@ -45,7 +45,7 @@ import {
   showSaveModal,
   stateMonitorFactory,
   subscribeWithScope,
-  unhashUrl,
+  unhashUrl
 } from '../kibana_services';
 
 const {
