+++ src/legacy/ui/public/chrome/api/sub_url_hooks.js
@@ -19,9 +19,8 @@
 
 import url from 'url';
 
-import {
-  unhashUrl,
-} from '../../state_management/state_hashing';
+import { unhashUrl } from '../../../../../plugins/kibana_utils/public';
+import { toastNotifications } from '../../notify/toasts';
 
 export function registerSubUrlHooks(angularModule, internals) {
   angularModule.run(($rootScope, Private, $location) => {
@@ -29,8 +28,14 @@ export function registerSubUrlHooks(angularModule, internals) {
 
     function updateSubUrls() {
       const urlWithHashes = window.location.href;
-      const urlWithStates = unhashUrl(urlWithHashes);
-      internals.trackPossibleSubUrl(urlWithStates);
+      let urlWithStates;
+      try {
+        urlWithStates = unhashUrl(urlWithHashes);
+      } catch (e) {
+        toastNotifications.addDanger(e.message);
+      }
+
+      internals.trackPossibleSubUrl(urlWithStates || urlWithHashes);
     }
 
     function onRouteChange($event) {
