+++ src/plugins/kibana_utils/public/index.ts
@@ -28,3 +28,5 @@ export * from './errors';
 export * from './field_mapping';
 export * from './storage';
 export * from './storage/hashed_item_store';
+export * from './state_management/state_hash';
+export * from './state_management/url';
