+++ src/plugins/kibana_utils/public/state_management/state_hash/index.ts
@@ -17,5 +17,4 @@
  * under the License.
  */
 
-export { hashUrl, unhashUrl, hashQuery, unhashQuery } from './hash_unhash_url';
-export { createStateHash, isStateHash } from './state_hash';
+export * from './state_hash';
