+++ src/plugins/kibana_utils/public/state_management/state_hash/state_hash.ts
@@ -18,7 +18,7 @@
  */
 
 import { Sha256 } from '../../../../../core/public/utils';
-import { hashedItemStore } from '../../../../../plugins/kibana_utils/public';
+import { hashedItemStore } from '../../storage/hashed_item_store';
 
 // This prefix is used to identify hash strings that have been encoded in the URL.
 const HASH_PREFIX = 'h@';
