+++ src/plugins/kibana_utils/public/state_management/url/index.ts
@@ -17,25 +17,4 @@
  * under the License.
  */
 
-import chrome from 'ui/chrome';
-import { hashUrl } from '../../../../plugins/kibana_utils/public';
-import uiRoutes from 'ui/routes';
-import { fatalError } from 'ui/notify';
-
-uiRoutes.enable();
-uiRoutes
-  .when('/', {
-    resolve: {
-      url: function (AppState, globalState, $window) {
-        const redirectUrl = chrome.getInjected('redirectUrl');
-        try {
-          const hashedUrl = hashUrl(redirectUrl);
-          const url = chrome.addBasePath(hashedUrl);
-
-          $window.location = url;
-        } catch (e) {
-          fatalError(e);
-        }
-      }
-    }
-  });
+export * from './hash_unhash_url';
