+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -676,6 +676,8 @@
     "kibana-react.savedObjects.saveModal.saveButtonLabel": "保存",
     "kibana-react.savedObjects.saveModal.saveTitle": "{objectType} を保存",
     "kibana-react.savedObjects.saveModal.titleLabel": "タイトル",
+    "kibana_utils.stateManagement.url.unableToRestoreUrlErrorMessage": "URL を完全に復元できません。共有機能を使用していることを確認してください。",
+    "kibana_utils.stateManagement.url.unableToStoreHistoryInSessionErrorMessage": "セッションがいっぱいで安全に削除できるアイテムが見つからないため、Kibana は履歴アイテムを保存できません。\n\nこれは大抵新規タブに移動することで解決されますが、より大きな問題が原因である可能性もあります。このメッセージが定期的に表示される場合は、{gitHubIssuesUrl} で問題を報告してください。",
     "inspector.closeButton": "インスペクターを閉じる",
     "inspector.reqTimestampDescription": "リクエストの開始が記録された時刻です",
     "inspector.reqTimestampKey": "リクエストのタイムスタンプ",
