+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -677,6 +677,8 @@
     "kibana-react.savedObjects.saveModal.saveButtonLabel": "保存",
     "kibana-react.savedObjects.saveModal.saveTitle": "保存 {objectType}",
     "kibana-react.savedObjects.saveModal.titleLabel": "标题",
+    "kibana_utils.stateManagement.url.unableToRestoreUrlErrorMessage": "无法完整还原 URL，确保使用共享功能。",
+    "kibana_utils.stateManagement.url.unableToStoreHistoryInSessionErrorMessage": "Kibana 无法将历史记录项存储在您的会话中，因为其已满，并且似乎没有任何可安全删除的项。\n\n通常可通过移至新的标签页来解决此问题，但这会导致更大的问题。如果您有规律地看到此消息，请在 {gitHubIssuesUrl} 提交问题。",
     "inspector.closeButton": "关闭检查器",
     "inspector.reqTimestampDescription": "记录请求启动的时间",
     "inspector.reqTimestampKey": "请求时间戳",
